pyproject.toml
[tool.poetry]
name = "mammon"
version = "0.1.0"
description = "Autonomous DeFi Yield Optimizer with x402 Integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.3.0"
anthropic = "^0.39.0"
cdp-sdk = "^1.33.2"
web3 = "^7.0.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
streamlit = "^1.39.0"
pandas = "^2.2.0"
requests = "^2.32.0"
aiohttp = "^3.10.0"
coinbase-agentkit = "^0.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-timeout = "^2.3.0"
black = "^24.0.0"
ruff = "^0.7.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src --cov-report=html --cov-report=term-missing --timeout=30"
timeout = 30

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
